latest/pyproject.toml
[project]
name = "systematically-improving-rag"
version = "0.1.0"
description = "Systematically Improve Your RAG System with our interactive notebooks!"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "accelerate==1.3.0",
    "anthropic>=0.40.0",
    "bertopic>=0.16.4",
    "braintrust>=0.0.177",
    "bs4>=0.0.2",
    "cohere>=5.11.3",
    "datasets>=3.1.0",
    "faker>=33.1.0",
    "thefuzz==0.22.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jsonref>=1.1.0",
    "lancedb>=0.16.0",
    "markdownify>=0.13.1",
    "matplotlib>=3.9.2",
    "modal>=0.65.38",
    "nbformat>=5.10.4",
    "open-clip-torch>=2.29.0",
    "openai>=1.57.0",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pydantic>=2.10.1",
    "python-levenshtein>=0.26.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.0",
    "instructor==1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=2.2.0,<3.0.0",
    "streamlit>=1.40.1",
    "streamlit-shortcuts>=0.1.9",
    "tantivy>=0.22.0",
    "transformers==4.42.2",
    "eval_type_backport==0.2.2",
    "pydantic-evals[logfire]>=0.1.3",
    "docling==2.30.0",
    "pdf2image==1.17.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.0",
    "nbconvert>=7.16.6",
    "typer>=0.15.4",
    "chromadb>=1.0.13",
    "turbopuffer>=0.5.8",
    "pylance>=0.30.0",
    "ruff>=0.12.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "tiktoken>=0.9.0",
    "diskcache>=5.6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
